Struct jsonrpc_minihttp_server::ServerBuilder

pub struct ServerBuilder<M: Metadata = (), S: Middleware<M> = NoopMiddleware> { /* fields omitted */ }

Convenient JSON-RPC HTTP Server builder.

Methods

impl<M: Metadata + Default, S: Middleware<M>> ServerBuilder<M, S>

pub fn new<T>(handler: T) -> Self where
T: Into<MetaIoHandler<M, S>>,

Creates new ServerBuilder for given IoHandler.

By default:

Server is not sending any CORS headers.

Server is validating Host header.

impl<M: Metadata, S: Middleware<M>> ServerBuilder<M, S>

pub fn with_meta_extractor<T, E>(handler: T, extractor: E) -> Self where
T: Into<MetaIoHandler<M, S>>,
E: MetaExtractor<M>,

Creates new ServerBuilder for given IoHandler and meta extractor.

By default:

Server is not sending any CORS headers.

Server is validating Host header.

pub fn threads(self, threads: usize) -> Self

Sets number of threads of the server to run. (not available for windows) Panics when set to 0.

pub fn cors(
self,
cors_domains: DomainsValidation<AccessControlAllowOrigin>
) -> Self

Configures a list of allowed CORS origins.

pub fn meta_extractor<T: MetaExtractor<M>>(self, extractor: T) -> Self

Configures metadata extractor

pub fn allow_only_bind_host(self) -> Self

Allow connections only with Host header set to binding address.

pub fn allowed_hosts(self, allowed_hosts: DomainsValidation<Host>) -> Self

Specify a list of valid Host headers. Binding address is allowed automatically.

pub fn start_http(self, addr: &SocketAddr) -> Result<Server>

Start this JSON-RPC HTTP server trying to bind to specified SocketAddr.